+++ src/tests/disabled-nodes.test.tsx
@@ -1,3 +1,4 @@
+// @ts-nocheck
 import React, { useState } from 'react';
 import '@testing-library/jest-dom';
 import { render, act, fireEvent, screen } from '@testing-library/react';
@@ -6,6 +7,7 @@ import {
   FocusNode,
   useSetFocus,
   useFocusStoreDangerously,
+  useProcessKey,
 } from '../index';
 import { warning } from '../utils/warning';
 
@@ -64,5 +66,68 @@ describe('disabled FocusNodes', () => {
     ]);
 
     expect(warning).toHaveBeenCalledTimes(0);
+    expect(console.error).toHaveBeenCalledTimes(0);
+  });
+
+  it('does not warn when a focus tree mounts with a disabled child (gh-78)', () => {
+    render(
+      <FocusRoot>
+        <FocusNode focusId="parent">
+          <FocusNode focusId="A" disabled />
+          <FocusNode focusId="B" />
+        </FocusNode>
+      </FocusRoot>
+    );
+
+    expect(console.error).toHaveBeenCalledTimes(0);
+    expect(warning).toHaveBeenCalledTimes(0);
+  });
+
+  it('disabled nodes work nicely with numeric defaultFocusChild (gh-104)', () => {
+    let focusStore;
+    let processKey;
+
+    function TestComponent() {
+      focusStore = useFocusStoreDangerously();
+      processKey = useProcessKey();
+
+      return (
+        <FocusNode focusId="testRoot" orientation="horizontal">
+          <FocusNode focusId="nodeB" data-testid="nodeB" />
+          <FocusNode
+            orientation="vertical"
+            focusId="nodeA"
+            data-testid="nodeA"
+            defaultFocusChild={() => {
+              return 2;
+            }}>
+            <FocusNode focusId="nodeA-A" data-testid="nodeA-B" disabled />
+            <FocusNode focusId="nodeA-B" data-testid="nodeA-B" />
+            <FocusNode focusId="nodeA-C" data-testid="nodeA-C" />
+          </FocusNode>
+        </FocusNode>
+      );
+    }
+
+    render(
+      <FocusRoot>
+        <TestComponent />
+      </FocusRoot>
+    );
+
+    let focusState = focusStore.getState();
+    expect(focusState.focusedNodeId).toEqual('nodeB');
+    expect(focusState.focusHierarchy).toEqual(['root', 'testRoot', 'nodeB']);
+
+    act(() => processKey.right());
+
+    focusState = focusStore.getState();
+    expect(focusState.focusedNodeId).toEqual('nodeA-C');
+    expect(focusState.focusHierarchy).toEqual([
+      'root',
+      'testRoot',
+      'nodeA',
+      'nodeA-C',
+    ]);
   });
 });