+++ src/tests/focus-node-events.test.tsx
@@ -1,7 +1,9 @@
+// @ts-nocheck
 import React from 'react';
 import '@testing-library/jest-dom';
 import { render, fireEvent, screen } from '@testing-library/react';
 import { FocusRoot, FocusNode, useFocusStoreDangerously } from '../index';
+import { warning } from '../utils/warning';
 
 describe('FocusNode Events', () => {
   describe('onMove', () => {
@@ -73,6 +75,9 @@ describe('FocusNode Events', () => {
           nextChildNode: focusStore.getState().nodes.nodeB,
         })
       );
+
+      expect(warning).toHaveBeenCalledTimes(0);
+      expect(console.error).toHaveBeenCalledTimes(0);
     });
 
     it('calls when given a keyCode', () => {
@@ -200,6 +205,9 @@ describe('FocusNode Events', () => {
       expect(nodeAOnBlurred.mock.calls.length).toBe(1);
       expect(nodeBOnFocused.mock.calls.length).toBe(1);
       expect(nodeBOnBlurred.mock.calls.length).toBe(0);
+
+      expect(warning).toHaveBeenCalledTimes(0);
+      expect(console.error).toHaveBeenCalledTimes(0);
     });
   });
 
@@ -239,6 +247,7 @@ describe('FocusNode Events', () => {
       fireEvent.keyDown(window, {
         code: 'ArrowRight',
         key: 'ArrowRight',
+        targetNode: focusStore.getState().nodes.nodeA,
       });
 
       nodeEl = screen.getByTestId('nodeA');
@@ -248,6 +257,9 @@ describe('FocusNode Events', () => {
       focusState = focusStore.getState();
       expect(focusState.focusedNodeId).toEqual('nodeA');
       expect(focusState.focusHierarchy).toEqual(['root', 'nodeA']);
+
+      expect(warning).toHaveBeenCalledTimes(0);
+      expect(console.error).toHaveBeenCalledTimes(0);
     });
 
     it('stopPropagation', () => {
@@ -289,6 +301,9 @@ describe('FocusNode Events', () => {
       focusState = focusStore.getState();
       expect(focusState.focusedNodeId).toEqual('nodeB');
       expect(focusState.focusHierarchy).toEqual(['root', 'nodeB']);
+
+      expect(warning).toHaveBeenCalledTimes(0);
+      expect(console.error).toHaveBeenCalledTimes(0);
     });
   });
 
@@ -346,6 +361,7 @@ describe('FocusNode Events', () => {
           key: 'select',
           isArrow: false,
           node: focusStore.getState().nodes.testRoot,
+          targetNode: preFocusedNodeA,
         })
       );
 
@@ -357,8 +373,12 @@ describe('FocusNode Events', () => {
           // active state. This allows you to, say, prevent default
           // the action.
           node: preFocusedNodeA,
+          targetNode: preFocusedNodeA,
         })
       );
+
+      expect(warning).toHaveBeenCalledTimes(0);
+      expect(console.error).toHaveBeenCalledTimes(0);
     });
   });
 
@@ -414,6 +434,7 @@ describe('FocusNode Events', () => {
           key: 'back',
           isArrow: false,
           node: focusStore.getState().nodes.testRoot,
+          targetNode: focusStore.getState().nodes.nodeA,
         })
       );
 
@@ -422,8 +443,12 @@ describe('FocusNode Events', () => {
           key: 'back',
           isArrow: false,
           node: focusStore.getState().nodes.nodeA,
+          targetNode: focusStore.getState().nodes.nodeA,
         })
       );
+
+      expect(warning).toHaveBeenCalledTimes(0);
+      expect(console.error).toHaveBeenCalledTimes(0);
     });
 
     it('calls it when Backspace is pressed', () => {